requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
asyncio>=3.4.3
pandas>=2.0.0
hashlib2>=1.0.0
subprocess32>=3.5.4
json5>=0.9.14
typing-extensions>=4.5.0
pydantic>=2.0.0
rich>=13.0.0